+++ src/Bridge/Symfony/DependencyInjection/DbToolsExtension.php
@@ -56,7 +56,7 @@ final class DbToolsExtension extends Extension
         if (\is_dir($defaultDirectory)) {
             $anonymizerPaths[] = '%kernel.project_dir%/src/Anonymizer';
         }
-        $anonymizerPaths[] = \realpath(\dirname(__DIR__)) . '/Anonymization/Anonymizer';
+        $anonymizerPaths[] = \realpath(\dirname(__DIR__, 3)) . '/Anonymization/Anonymizer';
 
         $container->setParameter('db_tools.anonymization.anonymizer.paths', $anonymizerPaths);
